:bug: Check if response is set

+++ events/messageDelete.js
@@ -3,6 +3,7 @@ import { Event } from '../structures/Event'
 export default class MessageDeleteEvent extends Event {
     async run (msg) {
         if (msg.command)
-            msg.response.delete()
+            if (msg.response)
+                msg.response.delete()
     }
 }
